% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate.R
\name{cate}
\alias{cate}
\title{Calculate categorical statistics from numerical vectors in related to a threshold}
\usage{
cate(
  model,
  observation,
  threshold,
  cutoff = NA,
  nobs = 8,
  rname,
  to.plot = F,
  col = "#4444bb",
  pch = 19,
  lty = 3,
  lcol = "#333333",
  verbose = T,
  ...
)
}
\arguments{
\item{model}{numeric vector with paired model data}

\item{observation}{numeric vector with paired observation data}

\item{threshold}{reference value}

\item{cutoff}{(optionally the maximum) valid value for observation}

\item{nobs}{minimum number of observations}

\item{rname}{row name}

\item{to.plot}{TRUE to plot a scatter-plot}

\item{col}{color for points}

\item{pch}{pch of points}

\item{lty}{lty of threshold lines}

\item{lcol}{col of threshold lines}

\item{verbose}{display additional information}

\item{...}{arguments passed to plot}
}
\value{
a data.frame including: Accuracy (A); Critical Success Index (CSI); Probability of Detection (POD); Bias(B); False Alarm Ratio (FAR); Heidke Skill Score (HSS); Pearce skill Score (PSS) in %.
}
\description{
Calculate traditional statistics from two numerical vectors in related to a threshold
}
\examples{
model <- 0.02 * 1:100
data  <- abs(rnorm(100,0.01))
cate(model = model, observation = data, threshold = 1,
     to.plot = TRUE, rname = 'example')

}
