% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva3dm-package.R
\docType{package}
\name{eva3dm-package}
\alias{eva3dm}
\alias{eva3dm-package}
\title{eva3dm: Evaluation of 3D Meteorological and Air Quality Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools for post-process, evaluate and visualize results from 3d Meteorological and Air Quality models against point observations (i.e. surface stations) and grid (i.e. satellite) observations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://schuch666.github.io/eva3dm/}
  \item Report bugs at \url{https://github.com/Schuch666/eva3dm/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Schuch \email{underschuch@gmail.com} (\href{https://orcid.org/0000-0001-5977-4519}{ORCID})

}
\keyword{internal}
