% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_surgical.R
\name{extract_surgical}
\alias{extract_surgical}
\title{Extract time series from a list model files using time UTC, latitude, longitude, and altitude (meters).}
\usage{
extract_surgical(
  filelist,
  point,
  vars,
  model = "WRF",
  tol = 60,
  boundary = 5,
  include_indices = TRUE,
  include_model = TRUE,
  include_distances = TRUE,
  filename,
  verbose = TRUE
)
}
\arguments{
\item{filelist}{list of files to be read}

\item{point}{data.frame with time (POSIXct in UTC), lat (numeric in degrees), lon (numeric in degrees), alt (numeric in meters)}

\item{vars}{variable names}

\item{model}{'WRF' (default) or 'UFS'}

\item{tol}{tolerance in seconds}

\item{boundary}{number of grid points that are considered}

\item{include_indices}{to include t,i,j,k indices}

\item{include_model}{to include model time, latitude, longitude and altitude}

\item{include_distances}{include the distance from the center of the grid to the observations}

\item{filename}{name of a file to save the dataframe}

\item{verbose}{display additional information}
}
\value{
data.frame
}
\description{
Read output from a list of model (WRF or UFS) and calculate the column of trace gases.
}
