% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva.R
\name{\%i\%}
\alias{\%i\%}
\title{Returns the common columns}
\usage{
x \%i\% y
}
\arguments{
\item{x}{data.frame}

\item{y}{data.frame or character string}
}
\description{
results of d01 in d02 style syntax
}
\note{
a message is always displayed to keep easy to track and debug issues (with the results and the evaluation process).
}
