% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.R
\name{interp}
\alias{interp}
\title{interpolation}
\usage{
interp(x, y, verbose = F)
}
\arguments{
\item{x}{rast to be interpolated}

\item{y}{target of the interpolation}

\item{verbose}{display additional information (not used)}
}
\description{
function to project and interpolate rast
}
