% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overlay.R
\name{overlay}
\alias{overlay}
\title{Plot rast (SpatRaster) object}
\usage{
overlay(
  p,
  z,
  col = hcl.colors(41, "Blue-Red"),
  lim = range(z, na.rm = TRUE),
  symmetry = TRUE,
  pch = 19,
  cex = 1,
  outside = TRUE,
  add = TRUE,
  plg = list(tic = "none", shrink = 1),
  pax = list(),
  expand = 1,
  ...
)
}
\arguments{
\item{p}{SpatVector points}

\item{z}{values to plot}

\item{col}{color}

\item{lim}{range of values for scale}

\item{symmetry}{calculate symmetrical scale}

\item{pch}{type of point}

\item{cex}{character expansion for the points}

\item{outside}{to include values outside range}

\item{add}{add to existing plot}

\item{plg}{list of parameters passed to terra::add_legend}

\item{pax}{list of parameters passed to graphics::axis}

\item{expand}{to extand the plot region}

\item{...}{arguments to be passing to terra::plot}
}
\description{
Custon plot for SpatRaster (terra R-package) object based on terra package
}
\examples{
masp  <- terra::vect(paste0(system.file("extdata",package="eval3dmodel"),
                           "/masp.shp"))
BR    <- terra::vect(paste0(system.file("extdata",package="eval3dmodel"),
                            "/BR.shp"))

p     <- readRDS(paste0(system.file("extdata",package="eval3dmodel"),
                        "/BR-AQ.Rds"))
p$id  <- row.names(p)
point <- terra::vect(p)
point$NMB <- 1:45 - 20

terra::plot(BR, main = 'add points')
terra::lines(BR)
terra::lines(masp, col = 'gray')
overlay(point,point$NMB,cex = 1.4)

overlay(point,point$NMB,cex = 1.4, add = F, main = 'main plot')
terra::lines(BR)
terra::lines(masp, col = 'gray')

}
