% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff.R
\name{plot_diff}
\alias{plot_diff}
\title{Plot rast (SpatRaster) object}
\usage{
plot_diff(
  a,
  b,
  col,
  lim_1 = NA,
  lim_2 = NA,
  file,
  w = 800,
  h = 1200,
  pt = 18,
  ...
)
}
\arguments{
\item{a}{SpatVector points}

\item{b}{values to plot}

\item{col}{color}

\item{lim_1}{range of values for scale}

\item{lim_2}{calculate symmetrical scale}

\item{w}{width for png function}

\item{h}{height for png function}

\item{pt}{fond size for png function}

\item{...}{arguments to be passing to terra::plot}

\item{save}{name to save a .png file instead of plot}
}
\description{
Custon plot for SpatRaster (terra R-package) object based on terra package
}
\examples{
print('oi')

}
