% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rast.R
\name{plot_rast}
\alias{plot_rast}
\title{Plot rast (SpatRaster) object}
\usage{
plot_rast(
  r,
  color,
  ncolor = 21,
  log = FALSE,
  proj = FALSE,
  plg = list(tic = "none", shrink = 0.98),
  pax = list(),
  axe = FALSE,
  grid = FALSE,
  latitude = FALSE,
  longitude = FALSE,
  add_grid = FALSE,
  add_range = TRUE,
  ...
)
}
\arguments{
\item{r}{raster}

\item{color}{color scale}

\item{ncolor}{number of colors}

\item{log}{TRUE to plot in log-scale}

\item{proj}{TRUE to project the raster to lat-lon}

\item{plg}{list of parameters passed to terra::add_legend}

\item{pax}{list of parameters passed to graphics::axis}

\item{axe}{plot axis}

\item{grid}{add grid}

\item{latitude}{add a latitude axis}

\item{longitude}{add a longitude axis}

\item{add_grid}{add a lat-lon grid (graticule style)}

\item{add_range}{add legend with max, average and min r values}

\item{...}{arguments to be passing to terra::plot}
}
\description{
Custon plot for SpatRaster (terra R-package) object based on terra package
}
\examples{
wrf <- paste(system.file("extdata", package = "eval3dmodel"),
                         "/wrfinput_d01", sep="")

r <- wrf_rast(file=wrf, name='XLAT')

plot_rast(r)

}
