% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q2rh.R
\name{q2rh}
\alias{q2rh}
\title{Convert absolute humidity to relative humidity}
\usage{
q2rh(q, t = 288.15, p = 101325)
}
\arguments{
\item{q}{vector of relative humidity (in Kg/Kg)}

\item{t}{vector of temperature (in K)}

\item{p}{pressure (in Pa)}
}
\description{
function to convert absolute humidity to relative humidity.
}
\note{
default values are from standard atmosphere (288.15 K / 101325 Pa)
}
\examples{
q2rh(q = 0.0002038, t = 302.45, p = 100800)

}
