% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain.R
\name{rain}
\alias{rain}
\title{conversion of model precipitation to hourly precipitation}
\usage{
rain(rainc, rainnc, verbose = TRUE)
}
\arguments{
\item{rainc}{data.frame or SpatRaster with RAINC variable}

\item{rainnc}{data.frame or SpatRaster with RAINNC variable}

\item{verbose}{set TRUE to display additional information}
}
\description{
function that convert model accumulated precipitation to hourly precipitation.
}
