% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_stats.R
\name{read_stat}
\alias{read_stat}
\title{Function to read stats and evaluation}
\usage{
read_stat(file, sep = ";", dec = ".", verbose = FALSE, ...)
}
\arguments{
\item{file}{model data.frame}

\item{sep}{the field separator string, passed to read.table function}

\item{dec}{he string to use for decimal points, passed to read.table function}

\item{verbose}{display additional information}

\item{...}{arguments passed to read.table functions}
}
\description{
Function to read stats and evaluation output
}
\examples{
sample <- read_stat(file    = paste0(system.file("extdata", package = "eva3dm"),"/sample.txt"),
                    verbose = TRUE)

sample <- read_stat(file    = paste0(system.file("extdata", package = "eva3dm"),"/sample.csv"),
                    verbose = TRUE)

}
