% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sate.R
\name{sate}
\alias{sate}
\title{Functions to model evaluation using satellite}
\usage{
sate(
  mo,
  ob,
  n = 6,
  min = NA,
  max = NA,
  rname,
  method = "bilinear",
  verbose = T,
  ...
)
}
\arguments{
\item{mo}{rast with model}

\item{ob}{rast with observations}

\item{n}{number of points from the boundary removed, default is 5}

\item{min}{minimum value cutoff}

\item{max}{maximum value cutoff}

\item{rname}{passed to stat}

\item{method}{passed to terra::resample}

\item{verbose}{set TRUE to display additional information}

\item{...}{other arguments passed to stat}
}
\value{
a data.frame
}
\description{
functions to evaluate the spatial performance using satellite
}
\note{
If a YOU DIED error message appears, means you are removing all the valid values using the arguments min or max.
}
