% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat}
\alias{stat}
\title{Calculate evaluation statistics from numerical vectors}
\usage{
stat(
  model,
  observation,
  wd = FALSE,
  cutoff = NA,
  cutoff_NME = NA,
  nobs = 8,
  rname,
  verbose = T,
  ...
)
}
\arguments{
\item{model}{numeric vector with paired model data}

\item{observation}{numeric vector with paired observation data}

\item{wd}{logical, set true to apply a rotation on wind direction, see notes}

\item{cutoff}{(optionally the maximum) valid value for observation}

\item{cutoff_NME}{(optionally the maximum) valid value for observation for NME, MFB and MFE}

\item{nobs}{minimum number of observations}

\item{rname}{row name}

\item{verbose}{display additional information}

\item{...}{extra arguments passed to scatter plot}
}
\value{
data.frame
}
\description{
Calculate statistical indexes (Number of pairs, observation average, model average, correlation, Index Of Agreement, Factor of 2, Root Mean Square Error, Mean Bias, Mean error, Gross Error, Mean Factor Bias, Mean Factor Error, Normalized Mean Bias, and Normalized Mean Bias) for model evaluation
}
\note{
the option wd = TRUE apply a rotation of 360 on model wind direction to minimize the angular difference.
}
\examples{
model <- 1:100
data  <- model + rnorm(100,0.2)
stat(model = model, observation = data)

}
