% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template}
\alias{template}
\title{Create templates for model evaluation}
\usage{
template(
  root = getwd(),
  template = "WRF",
  case = "case",
  env = "rspatial",
  scheduler = "SBATCH",
  partition = "main",
  project = "PROJECT",
  verbose = TRUE
)
}
\arguments{
\item{root}{directory to create the template}

\item{template}{template type (see notes)}

\item{case}{case to be evaluated}

\item{env}{name of the conda environment}

\item{scheduler}{scheduler name}

\item{partition}{name of the partition}

\item{project}{project name}

\item{verbose}{display additional information}
}
\description{
Create templates of code (r-scripts and job submission script) to read, post-process and evaluate model results.
}
\note{
templates available:\cr
 - WRF (model post-process for METAR + INMET)\cr
 - WRF-Chem (model post-process for METAR, AQS in Brazil and AERONET)\cr
 - IPEN (model post-process for one experimental site)\cr
 - METAR (download observations)\cr
 - MET (evaluation of meteorology)\cr
 - AQ (evaluation of air quality)\cr
 - GPCP (evaluation of precipitation using satellite)
}
\examples{
temp <- file.path(tempdir(),"POST")
template(root = temp,template = 'WRF', case = 'WRF-only')

}
