% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{vars}
\alias{vars}
\title{Function to return variable names}
\usage{
vars(file = NA, action = "get", verbose = FALSE)
}
\arguments{
\item{file}{file name}

\item{action}{'get' to return variable names or 'print' to print}

\item{verbose}{display additional information}
}
\value{
string
}
\description{
Return variable names of a NetCDF
}
\examples{
vars(paste0(system.file("extdata",package="eva3dm"),'/wrfinput_d01'))

}
