% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atr.R
\name{atr}
\alias{atr}
\title{Read and write attributes on a NetCDF file}
\usage{
atr(file = NA, var = "?", att = NA, action = "get", value = NA, verbose = F)
}
\arguments{
\item{file}{file name}

\item{var}{variable name, 0 to global and "?" to show options}

\item{att}{attribute names (NA for get all attnames)}

\item{action}{Print, write or get (return the value) an attribute}

\item{value}{value to write}

\item{verbose}{display additional information}
}
\description{
Read and write metadata information of a NetCDF files
}
\examples{
nc <- paste0(system.file("extdata",package="eval3dmodel"),'/wrfinput_d01')
atr(nc,0, action = 'print')
atr(nc,'Times', action = 'print')
atr(nc,'XLAT', action = 'print')
atr(nc,'XLONG', action = 'print')

atr(nc,'XLONG','MemoryOrder', action = 'print')
atr(nc,'XLONG','description', action = 'print')
atr(nc,'XLONG','units', action = 'print')
atr(nc,'XLONG','stagger', action = 'print')
atr(nc,'XLONG','FieldType', action = 'print')

}
