% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_column.R
\name{calculate_column}
\alias{calculate_column}
\title{Calculate column concentration of trace gases form WRF-Chem}
\usage{
calculate_column(
  file = file.choose(),
  name = NA,
  met = file,
  DU = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{file}{WRF output file}

\item{name}{trace gas name}

\item{met}{(optional) WRF output file for meteorological variables}

\item{DU}{true to change the units from 'molecules cm-1' to 'DU'}

\item{verbose}{display additional information}

\item{...}{extra arguments passed to ncdf4::ncvar_get}
}
\value{
SpatRaster object (from terra package).
}
\description{
Read output from WRF model and calculate the column of trace gases.
}
