% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate.R
\name{cate}
\alias{cate}
\title{Calculate evaluation statistics from numerical vectors}
\usage{
cate(
  model,
  observation,
  threshold,
  cutoff = NA,
  nobs = 8,
  rname,
  to.plot = F,
  verbose = T,
  ...
)
}
\arguments{
\item{model}{numeric vector with paired model data}

\item{observation}{numeric vector with paired observation data}

\item{threshold}{reference value}

\item{cutoff}{(optionally the maximum) valid value for observation}

\item{nobs}{minimum number of observations}

\item{rname}{row name}

\item{to.plot}{TRUE to plot a scatterplot}

\item{verbose}{display additional information}

\item{...}{}
}
\value{
data.frame
}
\description{
Calculate statistical indexes (Number of pairs, observation average, model average, correlation, Index Of Agreement, Factor of 2, Root Mean Square Error, Mean Bias, Mean error, Normalized Mean Bias, and Normalized Mean Bias) for model evaluation
}
\note{
the option wd = TRUE apply a rotation of 360 on model wind direction to minimize the angular difference.
}
\examples{
model <- 0.02 * 1:100
data  <- abs(rnorm(100,0.01))
cate(model = model, observation = data, threshold = 1,
     to.plot = TRUE, rname = 'example')

}
