% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{daily}
\alias{daily}
\title{Calculate daily mean, min or max}
\usage{
daily(data, time = "date", var, stat = mean, verbose = TRUE)
}
\arguments{
\item{data}{data.frame with time column and variable columns to be processed}

\item{time}{name of the time column (default is date) in POSIXct}

\item{var}{name of the columns to be calculated}

\item{stat}{function of the statistics to calculate (default is mean)}

\item{verbose}{display additional information}
}
\value{
data.frame
}
\description{
function to calculate Ozone Maximum Daily 8-hr Average or 8-hr moving Average for a data.frame
}
