% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva.R
\name{eva}
\alias{eva}
\title{Model statistical evaluation}
\usage{
eva(
  mo,
  ob,
  station = "ALL",
  wd = FALSE,
  fair = NULL,
  cutoff = NA,
  cutoff_NME = NA,
  no_tz = FALSE,
  nobs = 8,
  rname = station,
  eval_function = stat,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{mo}{data.frame with model data}

\item{ob}{data.frame with observation data}

\item{station}{name of the station or "ALL", see notes}

\item{wd}{default is FALSE, see notes}

\item{fair}{model data.frame (or list of names) to perform a fair comparison, see notes}

\item{cutoff}{minimum (optionally the maximum) valid value for observation}

\item{cutoff_NME}{minimum (optionally the maximum) valid value for observation for NME}

\item{no_tz}{ignore tz from input (force GMT)}

\item{nobs}{minimum number of valid observations, default is 8}

\item{rname}{name of the output row (default is station)}

\item{eval_function}{evaluation function (default is stat)}

\item{verbose}{display additional information}

\item{...}{arguments to be passing to stats and plot}
}
\description{
Statistical (or categorical) evaluation from 2 data.frames. The input data.frames (model and observation)
must contain a "date" column (containing POSIXlt). The function perform some simple case tests and
perform the time pairing of observations and model data and can calculate the statistical evaluation or
categorical evaluation.
}
\note{
for wind direction a rotation of 360 (or -360) is applied to minimize the wind direction difference.

If station == 'ALL' (default) all the columns from observations are combined in one column
(same for observation) and all the columns are evaluated together.

Special thanks to Kiarash and Libo to help to test the wind direction option.
}
\examples{
model <- readRDS(paste0(system.file("extdata",package="eval3dmodel"),
                        "/model.Rds"))
obs   <- readRDS(paste0(system.file("extdata",package="eval3dmodel"),
                        "/obs.Rds"))

# if there is no observed data
# the function return an empty row
table <- eva(mo = model, ob = obs, station = "VVIbes")
print(table)

# new tests with no data will be discated
table <- eva(mo = model, ob = obs, station = "VVIbes")
print(table)

# if the station are not in the input data frame a message is displayed
# and the function return an empty row
table <- eva(mo = model, ob = obs, station = "Ibirapuera")
print(table)

# calculating statistical with a few observed values
table <- eva(mo = model, ob = obs, station = "Americana")
print(table)

# calculating categorical (using 2 for threshold) with a few observed values
table <- eva(mo = model, ob = obs, station = "Americana",
             eval_function = cate, threshold = 2)
print(table)

# calculating categorical (using 2 for threshold) with a few observed values
table <- eva(mo = model, ob = obs, station = "Americana",
             eval_function = cate, threshold = 10)
print(table)

}
\references{
Mughal MO, Lynch M, Yu F, McGann B, Jeanneret F, Sutton J (2017)
Wind modeling, validation and sensitivity study using Weather
Research and Forecasting model in complex terrain. Environ
Model Softw 90:107–125. https://doi.org/10.1016/j.envsoft.2017.
01.009
}
