% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.R
\name{interp}
\alias{interp}
\title{Interpolation (project and resample)}
\usage{
interp(x, y, method = "bilinear", verbose = F)
}
\arguments{
\item{x}{rast to be interpolated}

\item{y}{target rast of the interpolation}

\item{method}{passed to terra::resample}

\item{verbose}{display additional information (not used)}
}
\description{
function to project and interpolate rast
}
