% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma8h.R
\name{ma8h}
\alias{ma8h}
\title{8-hour moving average}
\usage{
ma8h(data, time = "date", var, verbose = TRUE)
}
\arguments{
\item{data}{data.frame with time column and variable columns to be processed}

\item{time}{name of the time column (defoult is date) in POSIXct}

\item{var}{name of the columns to be calculated}

\item{verbose}{display additional information}
}
\value{
data.frame
}
\description{
function to calculate Ozone 8-hour moving average for a data.frame
}
