% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda8.R
\name{mda8}
\alias{mda8}
\title{Maximum Daily 8-hr Average}
\usage{
mda8(data, time = "date", var, verbose = TRUE)
}
\arguments{
\item{data}{data.frame with time column and variable columns to be processed}

\item{time}{name of the time column (defoult is date)}

\item{var}{name of the columns to be calculated}

\item{verbose}{display additional information}
}
\value{
data.frame
}
\description{
function to calculate Ozone Maximum Daily 8-hr Average or 8-hr moving Average for a data.frame
}
