% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff.R
\name{plot_diff}
\alias{plot_diff}
\title{Plot rast (SpatRaster) object}
\usage{
plot_diff(
  x,
  y,
  col,
  absolute = T,
  relative = T,
  lim_1 = NA,
  lim_2 = NA,
  units = c("", expression("\%")),
  w = 800,
  h = 1200,
  pt = 18,
  ...
)
}
\arguments{
\item{x}{SpatVector points}

\item{y}{values to plot}

\item{col}{color}

\item{absolute}{to plot absolute difference}

\item{relative}{to plot relative difference}

\item{lim_1}{range of values for scale}

\item{lim_2}{calculate symmetrical scale}

\item{units}{units annotation}

\item{w}{width for png function}

\item{h}{height for png function}

\item{pt}{fond size for png function}

\item{...}{arguments to be passing to terra::plot}
}
\description{
Custon diffrerence plots (x - y) of plot for SpatRaster (terra R-package) object based on terra package
}
\examples{
print('oi')

}
