% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff.R
\name{plot_diff}
\alias{plot_diff}
\title{Plot the difference from two rast (SpatRaster) object}
\usage{
plot_diff(
  x,
  y,
  col,
  absolute = T,
  relative = T,
  lim_1 = NA,
  lim_2 = NA,
  units = c("", expression("\%")),
  ...
)
}
\arguments{
\item{x}{SpatVector points}

\item{y}{values to plot}

\item{col}{color}

\item{absolute}{to plot absolute difference}

\item{relative}{to plot relative difference}

\item{lim_1}{range of values for scale}

\item{lim_2}{calculate symmetrical scale}

\item{units}{units annotation}

\item{...}{arguments to be passing to plot_raster}
}
\description{
Custon difference plots (x - y) of plot for SpatRaster (terra R-package) object based on terra package
}
\examples{

folder <- system.file("extdata",package="eva3dm")
wrf    <- paste0(folder,"/wrfinput_d01")
A      <- wrf_rast(wrf,'XLAT')
B      <- wrf_rast(wrf,'XLONG')
plot_diff(A,B,int = 2)
}
