% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rast.R
\name{plot_rast}
\alias{plot_rast}
\title{Plot rast (SpatRaster) object}
\usage{
plot_rast(
  r,
  color,
  ncolor = 21,
  proj = FALSE,
  plg = list(tic = "none", shrink = 0.98),
  pax = list(),
  latlon = FALSE,
  grid = TRUE,
  latitude = TRUE,
  longitude = TRUE,
  int = 10,
  grid_int = 10,
  add_range = TRUE,
  log = FALSE,
  hard_zlim = F,
  ...
)
}
\arguments{
\item{r}{raster}

\item{color}{color scale}

\item{ncolor}{number of colors}

\item{proj}{TRUE to project the raster to lat-lon}

\item{plg}{list of parameters passed to terra::add_legend}

\item{pax}{list of parameters passed to graphics::axis}

\item{latlon}{plot lat-lon axis (defoult), FALSE for r original axis}

\item{grid}{add grid (graticule style)}

\item{latitude}{add a latitude axis}

\item{longitude}{add a longitude axis}

\item{int}{interval of latitude and longitude lines}

\item{grid_int}{interval of grid lines}

\item{add_range}{add legend with max, average and min r values}

\item{log}{plot in log scale}

\item{hard_zlim}{hard lim}

\item{...}{arguments to be passing to terra::plot}
}
\description{
Custon plot for SpatRaster (terra R-package) object based on terra package
}
\examples{
wrf <- paste(system.file("extdata", package = "eval3dmodel"),
                         "/wrfinput_d01", sep="")

r <- wrf_rast(file=wrf, name='XLAT')

plot_rast(r)

}
