% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_to_netcdf.R
\name{rast_to_netcdf}
\alias{rast_to_netcdf}
\title{Function to convert (and save) a SpatRaster object to an array (and save on a Netcdf file)}
\usage{
rast_to_netcdf(r, file, name, verbose = TRUE)
}
\arguments{
\item{r}{SpatRaster object}

\item{file}{Netcdf file name}

\item{name}{variable name on a Netcdf file}

\item{verbose}{display additional information}
}
\description{
Conversion of SpatRaster to array and opcionally save on a Netcdf File.
}
\examples{
folder   <- system.file("extdata",package="eval3dmodel")
wrf_file <- paste0(folder,"/wrf.day1.o3.nc")

Rast     <- wrf_rast(wrf_file,'o3')
A        <- rast_to_netcdf(Rast)

}
