% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh2q.R
\name{rh2q}
\alias{rh2q}
\title{Convert relative humidity to absolute humidity}
\usage{
rh2q(rh, temp = 15)
}
\arguments{
\item{rh}{vector of relative humidity (in percentage)}

\item{temp}{vector of temperature (in celcius)}
}
\description{
function to convert humidity to absolute humidity using tetens formula, assuming standart atmosphere conditions.
}
\note{
default values are from standard atmosphere (288.15 K)
}
\examples{
rh2q(99)

}
