% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template}
\alias{template}
\title{Create templates for model evaluation}
\usage{
template(
  root,
  template = "WRF",
  case = "case",
  env = "rspatial",
  scheduler = "SBATCH",
  partition = "main",
  project = "PROJECT",
  verbose = TRUE
)
}
\arguments{
\item{root}{directory to create the template}

\item{template}{Character; One of  of the following:
\tabular{lllll}{
  \strong{argument}\tab \strong{TYPE}\tab \strong{MODEL}\tab \strong{OBSERVATION}\cr
  WRF\tab post-process \tab WRF \tab METAR and INMET\cr
  WRF-3\tab post-process \tab WRF triple nested \tab METAR\cr
  WRF-Chem\tab post-process \tab WRF-Chem \tab METAR, AQS in Brazil and AERONET\cr
  EXP\tab post-process \tab WRF-Chem \tab METAR, PBL variables and composition\cr
  CAMX\tab post-process \tab CAMx \tab AERONET \cr
  CAMX-3\tab post-process \tab CAMx triple nested \tab AERONET \cr
  NCO\tab post-process \tab WRF \tab Satellite\cr
  terra\tab post-process \tab WRF \tab Satellite\cr
  SAT\tab evaluation \tab WRF \tab Satellite (GPCP)\cr
  MET\tab evaluation \tab WRF \tab METAR\cr
  MET-3\tab evaluation \tab WRF triple nested \tab METAR\cr
  AQ\tab evaluation \tab WRF or CAMx \tab O3, Max O3 and PM2.5\cr
  AQS_BR\tab download \tab any \tab AQS in Brazil for SP and RJ\cr
  METAR\tab downlaad \tab any \tab METAR from ASOS\cr
  INMET\tab pre-processing \tab any \tab INMET (automatic and conventional)\cr
  merge\tab pre-processing \tab any \tab merge INMET (automatic and conventional)\cr
  ISD\tab pre-processing \tab any \tab METAR\cr
}}

\item{case}{case to be evaluated}

\item{env}{name of the conda environment}

\item{scheduler}{job scheduler used (SBATCH or PBS)}

\item{partition}{partition name}

\item{project}{project name}

\item{verbose}{display additional information}
}
\value{
no value returned, create folders and other template scripts
}
\description{
Create templates of code (r-scripts and bash job-submission script) to read, post-process and evaluate model results.
}
\examples{
temp <- file.path(tempdir(),"POST")
template(root = temp,template = 'WRF', case = 'WRF-only')

}
