% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template}
\alias{template}
\title{Create templates for meodel evaluation}
\usage{
template(
  template = "WRF-METAR",
  case = "WRF-only",
  partition = "zhang",
  env = "rspatial",
  root = getwd(),
  verbose = TRUE
)
}
\arguments{
\item{template}{template name}

\item{case}{case to be evaluated}

\item{partition}{name of the partition}

\item{env}{name of the conda environment}

\item{root}{directory to create the template}

\item{verbose}{display additional information}
}
\description{
Calculate traditional statistics from two numerical vectors in related to a threshold
}
\examples{
temp <- file.path(tempdir(),"POST")
template(root = temp)

}
