% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrf_rast.R
\name{wrf_rast}
\alias{wrf_rast}
\title{Creates rast object wrf file}
\usage{
wrf_rast(
  file = file.choose(),
  name = NA,
  map,
  level = 1,
  latlon = FALSE,
  as_polygons = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{file}{wrf file}

\item{name}{variable name}

\item{map}{(optional) file with lat-lon variables and grid information}

\item{level}{only for 4d data, default is 1 (surface)}

\item{latlon}{logical (default is FALSE), set TRUE project the output to "+proj=longlat +datum=WGS84 +no_defs"}

\item{as_polygons}{logical, true to return a poligon instead of a raster}

\item{verbose}{display additional information}

\item{...}{extra arguments passed to ncdf4::ncvar_get}
}
\description{
Creates a rast (terra R-package) object from a variable from wrf file (or another compatible NetCDF)
}
\examples{
{

wrf <- paste(system.file("extdata", package = "eval3dmodel"),
                         "/wrfinput_d01", sep="")

r <- wrf_rast(file=wrf, name='XLAT')

library(terra)
plot(r, axes = TRUE)
}
}
